+++ src/Console/Commands/MakeMiddlewareCommand.php
@@ -32,6 +32,7 @@ class MakeMiddlewareCommand extends AbstractCommand
     {
         $this->setName('make:middleware')->setDescription('Create a new PSR-15 middleware class');
         $this->addArgument('name', InputArgument::REQUIRED, 'Name of the class');
+
         $this->addOption('path', null, InputOption::VALUE_OPTIONAL, 'Path for the file to be created', 'src/Http/Middleware');
         $this->addOption('namespace', null, InputOption::VALUE_OPTIONAL, 'Namespace of the class', 'App\Http\Middleware');
         $this->addOption('package', null, InputOption::VALUE_OPTIONAL, 'Name of the package', 'App');