+++ src/Integrations/Illuminate/DatabaseIntegration.php
@@ -27,10 +27,11 @@ class DatabaseIntegration implements \Rougin\Slytherin\Integration\IntegrationIn
         $capsule = new \Illuminate\Database\Capsule\Manager;
 
         foreach ($config->get('database', array()) as $key => $value) {
-            if (is_array($value) === true) {
-                $value['collation'] = $config->get('database.' . $key . '.collation', 'utf8_unicode_ci');
-                $value['prefix'] = $config->get('database.' . $key . '.prefix', '');
+            $value['collation'] = $config->get('database.' . $key . '.collation', 'utf8_unicode_ci');
 
+            $value['prefix'] = $config->get('database.' . $key . '.prefix', '');
+
+            if (is_array($value) === true) {
                 $key = ($key === $config->get('database.default')) ? 'default' : $key;
 
                 $capsule->addConnection($value, $key);
@@ -38,7 +39,6 @@ class DatabaseIntegration implements \Rougin\Slytherin\Integration\IntegrationIn
         }
 
         $capsule->setAsGlobal();
-
         $capsule->bootEloquent();
 
         return $container->set(get_class($capsule), $capsule);